upgrades to result view

+++ app/api/results/[id]/route.ts
@@ -1,4 +1,4 @@
-import { deleteResult, getResult } from '@/app/lib/results/async';
+import { deleteResult, getCompleteResult, resultExists } from '@/app/lib/results/async';
 import { exportYAMLOrJSON } from '@/app/lib/results/helpers';
 import { NextRequest, NextResponse } from 'next/server';
 
@@ -8,10 +8,13 @@ export async function DELETE(request: NextRequest) {
 	if (duosmiumID === undefined) {
 		return new NextResponse('No result specified!', { status: 400 });
 	} else {
+		if ((await resultExists(duosmiumID)) === false) {
+			return new NextResponse('Result ' + duosmiumID + ' does not exist!', { status: 404 });
+		}
 		try {
 			await deleteResult(duosmiumID);
 		} catch (e) {
-			return new NextResponse('Result ' + duosmiumID + ' does not exist!', { status: 404 });
+			return new NextResponse('Server Error', { status: 500 });
 		}
 	}
 	return new NextResponse(null, { status: 204 });
@@ -23,23 +26,26 @@ export async function GET(request: NextRequest) {
 	if (duosmiumID === undefined) {
 		return new NextResponse('No result specified!', { status: 400 });
 	} else {
+		if ((await resultExists(duosmiumID)) === false) {
+			return new NextResponse('Result ' + duosmiumID + ' does not exist!', { status: 404 });
+		}
 		try {
-			const result = await getResult(duosmiumID);
+			const result = await getCompleteResult(duosmiumID);
 			return exportYAMLOrJSON(url, result, duosmiumID);
 		} catch (e) {
-			return new NextResponse('Result ' + duosmiumID + ' does not exist!', { status: 404 });
+			return new NextResponse('Server Error', { status: 500 });
 		}
 	}
 }
 
 export async function PATCH() {
-	return new NextResponse(null, { status: 501 });
+	return new NextResponse(null, { status: 405 });
 }
 
 export async function POST() {
-	return new NextResponse(null, { status: 501 });
+	return new NextResponse(null, { status: 405 });
 }
 
 export async function PUT() {
-	return new NextResponse(null, { status: 501 });
+	return new NextResponse(null, { status: 405 });
 }
